changed af from decimal to double in variant

+++ src/test/scala/org/kidsfirstdrc/dwh/testutils/external/TopmedBravoOutput.scala
@@ -7,11 +7,11 @@ case class TopmedBravoOutput(chromosome: String = "2",
                              reference: String = "T",
                              alternate: String = "A",
                              ac: Int = 5,
-                             af: BigDecimal = 0.5,
+                             af: Double = 0.5,
                              an: Int = 10,
                              homozygotes: Int = 5,
                              heterozygotes: Int = 0,
-                             qual: BigDecimal = 0.8,
+                             qual: Double = 0.8,
                              filters: Array[String] = Array(),
                              qual_filter: String = "")
